Add StyleCop and FxCop Analyzers
