+++ src/Samples/packages.config
@@ -1,5 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="Extended.Wpf.Toolkit" version="3.5.0" targetFramework="net45" />
+  <package id="Microsoft.CodeAnalysis.FxCopAnalyzers" version="2.9.3" targetFramework="net48" />
+  <package id="Microsoft.CodeAnalysis.VersionCheckAnalyzer" version="2.9.3" targetFramework="net48" />
+  <package id="Microsoft.CodeQuality.Analyzers" version="2.9.3" targetFramework="net48" />
+  <package id="Microsoft.NetCore.Analyzers" version="2.9.3" targetFramework="net48" />
+  <package id="Microsoft.NetFramework.Analyzers" version="2.9.3" targetFramework="net48" />
   <package id="Microsoft.Xaml.Behaviors.Wpf" version="1.0.1" targetFramework="net45" />
+  <package id="StyleCop.Analyzers" version="1.1.118" targetFramework="net48" developmentDependency="true" />
 </packages>
