+++ src/SaneDevelopment.WPF.Controls.sln
@@ -7,6 +7,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SaneDevelopment.WPF.Control
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Samples", "Samples\Samples.csproj", "{FAC6E5E8-3F26-47B9-8936-09528A72FE81}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{3AF1EE6C-8E5D-4694-8BB6-9012FA4B24F5}"
+	ProjectSection(SolutionItems) = preProject
+		SaneDevelopment.WPF.Controls.ruleset = SaneDevelopment.WPF.Controls.ruleset
+	EndProjectSection
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
